explicit casting in derivative() (C++11 version); added using namespace for ADL

+++ include/boost/math/differentiation/autodiff_cpp11.hpp
@@ -91,7 +91,7 @@ template<typename... Orders>
 get_type_at<fvar<RealType,Order>,sizeof...(Orders)> fvar<RealType,Order>::derivative(Orders... orders) const
 {
     static_assert(sizeof...(Orders) <= depth, "Number of parameters to derivative(...) cannot exceed fvar::depth.");
-    return at(orders...) * product(boost::math::factorial<root_type>(orders)...);
+    return at(static_cast<std::size_t>(orders)...) * product(boost::math::factorial<root_type>(static_cast<unsigned>(orders))...);
 }
 
 template<typename RealType, size_t Order>
